Add continous td and support for continous environments. Add Q_best_action fallback when A(s) is not available. WIP

+++ GPI.h
@@ -6,13 +6,13 @@ template <typename State, typename Action>
 class GPI : public MDPSolver<State, Action> {
    protected:
     double m_discount_rate;
-    double m_policy_threshold;
+    long double m_policy_threshold;
 
     virtual void policy_evaluation(){};
     virtual bool policy_improvement() { return false; };
 
    public:
-    GPI(MDP<State, Action>& mdp_core, const double discount_rate, const double policy_threshold)
+    GPI(MDP<State, Action>& mdp_core, const double discount_rate, const long double policy_threshold)
         : MDPSolver<State, Action>(mdp_core), m_discount_rate(discount_rate), m_policy_threshold(policy_threshold) {}
 
     virtual void policy_iteration() {
